refactor(PatternHopFactory): rename and move. response to PR #2848.
