+++ src/main/java/org/opentripplanner/graph_builder/module/geometry/ShapeSegmentKey.java
@@ -3,7 +3,7 @@
 /**
  * 
  */
-package org.opentripplanner.routing.edgetype.factory;
+package org.opentripplanner.graph_builder.module.geometry;
 
 import org.opentripplanner.model.FeedScopedId;
