+++ src/main/java/org/opentripplanner/gtfs/GenerateTripPatternsOperation.java
@@ -13,7 +13,7 @@ import org.opentripplanner.model.StopTime;
 import org.opentripplanner.model.Trip;
 import org.opentripplanner.model.TripPattern;
 import org.opentripplanner.model.impl.OtpTransitServiceBuilder;
-import org.opentripplanner.routing.edgetype.factory.PatternHopFactory;
+import org.opentripplanner.graph_builder.module.geometry.GeometryAndBlockProcessor;
 import org.opentripplanner.routing.graph.AddBuilderAnnotation;
 import org.opentripplanner.routing.trippattern.Deduplicator;
 import org.opentripplanner.routing.trippattern.FrequencyEntry;
@@ -27,7 +27,7 @@ import java.util.Set;
 
 /**
  * This class is responsible for generating trip patterns when loading GTFS data. This was
- * previously done in the {@link PatternHopFactory},
+ * previously done in the {@link GeometryAndBlockProcessor},
  * and the code is extracted out of it to make the PatternHopFactory reusable
  * for NETEX and GTFS file import.
  */
