+++ src/main/java/org/opentripplanner/gtfs/RepairStopTimesForEachTripOperation.java
@@ -12,7 +12,7 @@ import org.opentripplanner.graph_builder.annotation.RepeatedStops;
 import org.opentripplanner.model.StopTime;
 import org.opentripplanner.model.Trip;
 import org.opentripplanner.model.TripStopTimes;
-import org.opentripplanner.routing.edgetype.factory.PatternHopFactory;
+import org.opentripplanner.graph_builder.module.geometry.GeometryAndBlockProcessor;
 import org.opentripplanner.routing.graph.AddBuilderAnnotation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -23,8 +23,8 @@ import java.util.List;
 
 /**
  * This class is responsible for cleaning stop times, removing duplicates, correcting bad data
- * and so on. This was previously done in the {@link PatternHopFactory}. The code is
- * extracted out of {@link PatternHopFactory} to make the it reusable for the NeTEx and GTFS file
+ * and so on. This was previously done in the {@link GeometryAndBlockProcessor}. The code is
+ * extracted out of {@link GeometryAndBlockProcessor} to make the it reusable for the NeTEx and GTFS file
  * import. This do only apply to GTFS imports.
  */
 public class RepairStopTimesForEachTripOperation {
