+++ src/test/java/org/opentripplanner/ConstantsForTests.java
@@ -13,7 +13,7 @@ import org.opentripplanner.netex.configure.NetexConfig;
 import org.opentripplanner.netex.loader.NetexBundle;
 import org.opentripplanner.openstreetmap.impl.AnyFileBasedOpenStreetMapProviderImpl;
 import org.opentripplanner.openstreetmap.services.OpenStreetMapProvider;
-import org.opentripplanner.routing.edgetype.factory.PatternHopFactory;
+import org.opentripplanner.graph_builder.module.geometry.GeometryAndBlockProcessor;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.standalone.GraphBuilderParameters;
 import org.opentripplanner.standalone.Router;
@@ -111,7 +111,7 @@ public class ConstantsForTests {
                         .withGraphBuilderAnnotationsAndDeduplicator(portlandGraph)
                         .build();
                 AddTransitModelEntitiesToGraph.addToGraph(portlandContext, portlandGraph);
-                PatternHopFactory factory = new PatternHopFactory(portlandContext);
+                GeometryAndBlockProcessor factory = new GeometryAndBlockProcessor(portlandContext);
                 factory.run(portlandGraph);
             }
             // Link transit stops to streets
@@ -172,7 +172,7 @@ public class ConstantsForTests {
         }
         AddTransitModelEntitiesToGraph.addToGraph(context, graph);
 
-        PatternHopFactory factory = new PatternHopFactory(context);
+        GeometryAndBlockProcessor factory = new GeometryAndBlockProcessor(context);
         factory.run(graph);
         graph.putService(
                 CalendarServiceData.class, context.getCalendarServiceData()
