+++ src/test/java/org/opentripplanner/routing/algorithm/TestAStar.java
@@ -7,7 +7,7 @@ import org.opentripplanner.gtfs.GtfsContext;
 import org.opentripplanner.model.calendar.CalendarServiceData;
 import org.opentripplanner.routing.algorithm.astar.AStar;
 import org.opentripplanner.routing.core.RoutingRequest;
-import org.opentripplanner.routing.edgetype.factory.PatternHopFactory;
+import org.opentripplanner.graph_builder.module.geometry.GeometryAndBlockProcessor;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.Vertex;
 import org.opentripplanner.routing.spt.GraphPath;
@@ -29,7 +29,7 @@ public class TestAStar extends TestCase {
         GtfsContext context = contextBuilder(ConstantsForTests.CALTRAIN_GTFS).build();
 
         Graph gg = new Graph();
-        PatternHopFactory factory = new PatternHopFactory(context);
+        GeometryAndBlockProcessor factory = new GeometryAndBlockProcessor(context);
         factory.run(gg);
         gg.putService(
                 CalendarServiceData.class, context.getCalendarServiceData()
