+++ src/test/java/org/opentripplanner/routing/algorithm/TestFares.java
@@ -13,7 +13,7 @@ import org.opentripplanner.routing.core.FareComponent;
 import org.opentripplanner.routing.core.Money;
 import org.opentripplanner.routing.core.RoutingRequest;
 import org.opentripplanner.routing.core.WrappedCurrency;
-import org.opentripplanner.routing.edgetype.factory.PatternHopFactory;
+import org.opentripplanner.graph_builder.module.geometry.GeometryAndBlockProcessor;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.impl.SeattleFareServiceFactory;
 import org.opentripplanner.routing.services.FareService;
@@ -37,7 +37,7 @@ public class TestFares extends TestCase {
 
         Graph gg = new Graph();
         GtfsContext context = contextBuilder(ConstantsForTests.CALTRAIN_GTFS).build();
-        PatternHopFactory factory = new PatternHopFactory(context);
+        GeometryAndBlockProcessor factory = new GeometryAndBlockProcessor(context);
         factory.run(gg);
         gg.putService(
                 CalendarServiceData.class, context.getCalendarServiceData()
@@ -113,7 +113,7 @@ public class TestFares extends TestCase {
     	Graph gg = new Graph();
         GtfsContext context = contextBuilder(ConstantsForTests.KCM_GTFS).build();
         
-        PatternHopFactory factory = new PatternHopFactory(context);
+        GeometryAndBlockProcessor factory = new GeometryAndBlockProcessor(context);
         factory.setFareServiceFactory(new SeattleFareServiceFactory());
         
         factory.run(gg);
@@ -152,7 +152,7 @@ public class TestFares extends TestCase {
     public void testFareComponent() throws Exception {
         Graph gg = new Graph();
         GtfsContext context = contextBuilder(ConstantsForTests.FARE_COMPONENT_GTFS).build();
-        PatternHopFactory factory = new PatternHopFactory(context);
+        GeometryAndBlockProcessor factory = new GeometryAndBlockProcessor(context);
         factory.run(gg);
         gg.putService(
                 CalendarServiceData.class, context.getCalendarServiceData()
