+++ src/test/java/org/opentripplanner/routing/algorithm/TestGraphPath.java
@@ -9,7 +9,7 @@ import org.opentripplanner.model.calendar.CalendarServiceData;
 import org.opentripplanner.routing.algorithm.astar.AStar;
 import org.opentripplanner.routing.core.RoutingRequest;
 import org.opentripplanner.routing.core.State;
-import org.opentripplanner.routing.edgetype.factory.PatternHopFactory;
+import org.opentripplanner.graph_builder.module.geometry.GeometryAndBlockProcessor;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.Vertex;
 import org.opentripplanner.routing.spt.GraphPath;
@@ -34,7 +34,7 @@ public class TestGraphPath extends TestCase {
     public void setUp() throws Exception {
         GtfsContext context = contextBuilder(ConstantsForTests.FAKE_GTFS).build();
         graph = new Graph();
-        PatternHopFactory hl = new PatternHopFactory(context);
+        GeometryAndBlockProcessor hl = new GeometryAndBlockProcessor(context);
         hl.run(graph);
         graph.putService(
                 CalendarServiceData.class, context.getCalendarServiceData()
